third_party/WebKit/LayoutTests/editing/spelling/design-mode-spellcheck-off.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body spellcheck="false">
asdf is misspelled.
<script>
description("This tests whether WebKit does not spell check in 'designMode' "
    + "when spellcheck='false'. To test manually, click 'asdf' above. "
    + "There should be no spelling marker for 'asdf' after the click.");

jsTestIsAsync = true;

if (window.internals) {
    internals.settings.setUnifiedTextCheckerEnabled(true);
    internals.settings.setAsynchronousSpellCheckingEnabled(true);
}

document.designMode = "on";

document.body.addEventListener('mousedown', function() {
    if (document.activeElement != document.body) {
        testFailed('Active element after click is not body element');
        finishJSTest();
    }

    if (window.internals)
        shouldBecomeEqual('internals.hasSpellingMarker(document, 0, 4)', 'false', finishJSTest);
});

if (window.eventSender) {
    var x = document.body.offsetLeft + 10;
    var y = document.body.offsetTop + 17;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

</script> 
</body>
</html>
